seeded the trip data - learned db:reset is buggy - might not validate your input and might duplicate instances - trip cost is in cents - need to generate new migration to change col name

+++ db/migrate/20170404214217_create_trips.rb
@@ -5,7 +5,7 @@ class CreateTrips < ActiveRecord::Migration[5.0]
       t.integer :passenger_id
       t.date :date
       t.integer :rating
-      t.decimal :cost
+      t.integer :cost
       t.belongs_to :driver
       t.belongs_to :passenger
       t.timestamps
